# Perf: Defer SARIF log source map generation.