This is an alternative to #332. In #322 we selectively disable. In this one we always defer. This is in response to feedback that not all large SARIF files exclusively enter via the API.

Addresses a perf issue where a 150MB log takes 14s to load. With this fix it takes 3.5s. Memory pressure is no longer limited at 500MB. It seems to be smooth over 1.5GB now (no heap issues).

Note: Clicking on a "navigate to log result source" will be slower the first time since we are deferring the work. However, it might not be perceptible and also this is a producer-targeted feature (narrow audience). Thus we can deal with adding a progress bar at a later date (upon feedback).

#### Attaching our internal mail thread here:

I investigated (profiled) Roy's 150MB SARIF log. On my (slow) machine it takes ~14s to load. 10.5s of that is generating the source map. We can (situationally) drop the source map and shave 75% of the time.

What is the source map for? When you click the "6101_1.sarif" link below, it takes to you the region in the SARIF Log source. The source map powers that. A feature primarily for SARIF producers, not consumers.

![Screen Shot 2020-08-14 at 2 02 16 AM](https://user-images.githubusercontent.com/3452831/90277867-37cebb80-de1b-11ea-80b3-db853c32bcb7.png)

A few options to situationally disable the source map:
* Auto-disable on large files. Not great as many small files can still cause trouble.
* Add a user setting. I don't think it will be discoverable and we would still need smart defaults.
* Lazy-create the source map when that link is clicked. User may be surprised there is a long wait. Also makes the code more-stateful, which I recommend against.
* Auto-disable on API-triggered SARIF log opens. I'm recommending this one.

Thoughts? If nobody cares (say in 24 hours), I'll go with my recommendation. Thanks.